refactor: extract event_handler.rs module (Phase 1)
Successfully extracted X11 event handling into separate module while
maintaining all functionality. This is the first phase of incremental
SRP refactoring.

Changes:
- Created event_handler.rs with all handle_* methods
- Made necessary fields/methods pub(crate) for module access
- All 57 tests pass ✅
- All functionality verified working ✅

Next: Extract window_operations.rs module
