+++ src/event_handler.rs
@@ -1,0 +1,199 @@
+//! X11 event handling
+//!
+//! This module handles all X11 event processing and dispatching.
+
+use anyhow::Result;
+use std::process::Command;
+use tracing::{error, info};
+use x11rb::protocol::Event;
+use x11rb::protocol::xproto::*;
+
+use crate::WindowManager;
+
+impl<C: x11rb::connection::Connection> WindowManager<C> {
+    /// Main event dispatcher
+    pub(crate) fn handle_event(&mut self, event: Event) -> Result<()> {
+        match event {
+            Event::KeyPress(ev) => self.handle_key_press(ev),
+            Event::MapRequest(ev) => self.handle_map_request(ev),
+            Event::UnmapNotify(ev) => self.handle_unmap_notify(ev),
+            Event::ConfigureRequest(ev) => self.handle_configure_request(ev),
+            Event::DestroyNotify(ev) => self.handle_destroy_notify(ev),
+            Event::FocusIn(ev) => self.handle_focus_in(ev),
+            Event::FocusOut(ev) => self.handle_focus_out(ev),
+            Event::EnterNotify(ev) => self.handle_enter_notify(ev),
+            _ => {
+                #[cfg(debug_assertions)]
+                tracing::debug!("Unhandled event: {:?}", event);
+                Ok(())
+            }
+        }
+    }
+
+    /// Handles key press events
+    fn handle_key_press(&mut self, event: KeyPressEvent) -> Result<()> {
+        if let Some(command) = self.keyboard_manager.handle_key_press(&event) {
+            info!("Shortcut pressed, executing: {}", command);
+
+            // Handle window management commands
+            match command {
+                "focus_next" => return self.focus_next(),
+                "focus_prev" => return self.focus_prev(),
+                "swap_window_next" => return self.swap_window_next(),
+                "swap_window_prev" => return self.swap_window_prev(),
+                "destroy_window" => return self.destroy_focused_window(),
+                "toggle_fullscreen" => return self.toggle_fullscreen(),
+                "rotate_windows" => return self.rotate_windows(),
+                _ => {
+                    // Handle regular application commands
+                    let parts: Vec<&str> = command.split_whitespace().collect();
+                    if let Some(program) = parts.first() {
+                        let mut cmd = Command::new(program);
+
+                        // Add arguments if any
+                        if parts.len() > 1 {
+                            cmd.args(&parts[1..]);
+                        }
+
+                        // Set display environment
+                        cmd.env("DISPLAY", self.config.default_display());
+
+                        match cmd.spawn() {
+                            Ok(_) => info!("Successfully launched: {}", command),
+                            Err(e) => error!("Failed to launch {}: {}", command, e),
+                        }
+                    }
+                }
+            }
+        }
+        Ok(())
+    }
+
+    /// Handles window map requests
+    fn handle_map_request(&mut self, event: MapRequestEvent) -> Result<()> {
+        let window = event.window;
+        info!("Mapping window: {:?}", window);
+
+        // Set initial border attributes before mapping
+        self.configure_window_border(window, self.config.unfocused_border_color())?;
+
+        // Map the window
+        self.conn.map_window(window)?;
+
+        // Add to managed windows
+        self.add_window_to_layout(window);
+
+        // Set focus to new window
+        self.set_focus(window)?;
+
+        // Apply layout
+        self.apply_layout()?;
+
+        Ok(())
+    }
+
+    /// Handles window unmap notifications
+    fn handle_unmap_notify(&mut self, event: UnmapNotifyEvent) -> Result<()> {
+        let window = event.window;
+        info!("Unmapping window: {:?}", window);
+
+        // Check if this was intentionally unmapped (during fullscreen)
+        if self.intentionally_unmapped.contains(&window) {
+            info!("Window {:?} was intentionally unmapped, ignoring", window);
+            return Ok(());
+        }
+
+        // Only remove from managed windows if NOT intentionally unmapped
+        info!(
+            "Window {:?} closed by user, removing from management",
+            window
+        );
+        self.remove_window_from_layout(window);
+
+        // Update focus if focused window was unmapped
+        if self.focused_window == Some(window) {
+            // Focus first remaining window in BSP tree order
+            self.focused_window = self.get_first_window();
+            if let Some(next_focus) = self.focused_window {
+                self.set_focus(next_focus)?;
+            }
+        }
+
+        // Reapply layout
+        self.apply_layout()?;
+
+        Ok(())
+    }
+
+    /// Handles window configure requests
+    fn handle_configure_request(&mut self, event: ConfigureRequestEvent) -> Result<()> {
+        #[cfg(debug_assertions)]
+        tracing::debug!("Configure request for window: {:?}", event.window);
+
+        // For now, just honor the request
+        // In the future, we might want to be more selective
+        let values = ConfigureWindowAux::from_configure_request(&event);
+        self.conn.configure_window(event.window, &values)?;
+
+        Ok(())
+    }
+
+    /// Handles window destroy notifications
+    fn handle_destroy_notify(&mut self, event: DestroyNotifyEvent) -> Result<()> {
+        let window = event.window;
+        info!("Window destroyed: {:?}", window);
+
+        // Remove from managed windows
+        self.remove_window_from_layout(window);
+
+        // Clean up intentionally unmapped set to prevent memory leaks
+        self.intentionally_unmapped.remove(&window);
+
+        // Clear fullscreen if fullscreen window was destroyed
+        if self.fullscreen_window == Some(window) {
+            info!("Fullscreen window destroyed, exiting fullscreen mode");
+            self.fullscreen_window = None;
+        }
+
+        // Update focus if focused window was destroyed
+        if self.focused_window == Some(window) {
+            // Focus first remaining window in BSP tree order
+            self.focused_window = self.get_first_window();
+            if let Some(next_focus) = self.focused_window {
+                self.set_focus(next_focus)?;
+            }
+        }
+
+        // Reapply layout
+        self.apply_layout()?;
+
+        Ok(())
+    }
+
+    /// Handles focus in events
+    fn handle_focus_in(&mut self, _event: FocusInEvent) -> Result<()> {
+        #[cfg(debug_assertions)]
+        tracing::debug!("Focus in event for window: {:?}", _event.event);
+        Ok(())
+    }
+
+    /// Handles focus out events
+    fn handle_focus_out(&mut self, _event: FocusOutEvent) -> Result<()> {
+        #[cfg(debug_assertions)]
+        tracing::debug!("Focus out event for window: {:?}", _event.event);
+        Ok(())
+    }
+
+    /// Handles mouse enter events
+    fn handle_enter_notify(&mut self, event: EnterNotifyEvent) -> Result<()> {
+        let window = event.event;
+        #[cfg(debug_assertions)]
+        tracing::debug!("Mouse entered window: {:?}", window);
+
+        // Only focus if it's a managed window
+        if self.has_window(window) {
+            self.set_focus(window)?;
+        }
+        Ok(())
+    }
+}
